+++ tests/browser/pages/sso/sso_common.py
