+++ tests/browser/pages/sso/sso_confirm_your_email.py
