+++ tests/browser/pages/sso/sso_registration.py
