+++ tests/browser/pages/sso/sso_registration_confirmation.py
