+++ tests/browser/registry/pages.py
@@ -25,12 +25,9 @@ from pages.exread import (
 from pages import (
     international,
     soo_home,
-    sso_profile_about,
-    sso_registration,
-    sso_registration_confirmation,
-    sso_sign_in,
-    sso_sign_out,
 )
+from pages.sso import sso_sign_in, sso_registration, \
+    sso_registration_confirmation, sso_profile_about, sso_sign_out
 from pages.invest import invest_footer, invest_feedback, invest_contact_us, \
     invest_guide, invest_header, invest_industries, invest_industry, \
     invest_home
